GetAssetsInRepairsService to use Long instead of Integer

+++ backend/vaultops/src/main/java/com/vaultops/services/stats/GetAssetsInRepairsService.java
@@ -7,7 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
-public class GetAssetsInRepairsService implements Query<Void, Integer> {
+public class GetAssetsInRepairsService implements Query<Void, Long> {
 
     private final AssetRepository assetRepository;
 
@@ -16,8 +16,8 @@ public class GetAssetsInRepairsService implements Query<Void, Integer> {
     }
 
     @Override
-    public ResponseEntity<Integer> execute(Void input) {
-        Integer count = assetRepository.countAssetsByUsageStatus(Usage.SERVICE);
+    public ResponseEntity<Long> execute(Void input) {
+        Long count = assetRepository.countAssetsByUsageStatus(Usage.SERVICE);
         return ResponseEntity.ok(count);
     }
 }
